backend/pyproject.toml
[tool.poetry]
name = "aiapp"
version = "0.1.0"
description = ""
authors = ["He-llo-Wo-rld <serhiiforwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
django-axes = "^8.0.0"
requests = "^2.32.4"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
